Refactor : 게임실행에 관련된 이벤트리스너 로직 -> 파일로 분리(playGame.js)
